spi-nor: add a NO-OP flag for flash part
A NO-OP flag means the flash part selected can't be used and user should
select another one.

This happens if two or more flash parts share the same JEDEC ID but are
mutually incompatible with each other. A meta part must be defined with the
NO-OP flag for this situation to require the user to select another matched
part.

Signed-off-by: Weijie Gao <hackpascal@gmail.com>