Rename controller-related API names
Since the generic plugin framework was added, the concept of driver and
device can be reused by other plugins. It's necessary to rename the
controller-related API names not to use the 'driver' and 'device' keywords
directly. Just add namespace for these APIs.

This patch introduces API changes which makes the whole project plugins
incompatible with previous versions.

Signed-off-by: Weijie Gao <hackpascal@gmail.com>