controller: add support for serprog protocol
Add support for serprog protocol. This is designed as a controller plugin.
To use this plugin, user must write a device description json file to
specify which serial port device to use, and how to configure this device,
e.g. baudrate.

An example can be found in static/examples/serprog.json

Signed-off-by: Weijie Gao <hackpascal@gmail.com>