flash: spi-nor: add ability for reading multi-die flash using linear adressing
Some flashes use linear addressing to get rid of selecting die before
accessing. However there still exists the limit that only one die is active
for reading continuously, and the wrap around happens in the active die, not
the entire flash. So a single read command cannot read out data of the
entire flash.

To fix this, we have to limit the read size to make sure a single read
command cannot go across dir boundaries.

Signed-off-by: Weijie Gao <hackpascal@gmail.com>